Sign Up, Sign In, and Profile adjustments
Sign Up feature was implemented and is able to log user information in firebase database, along with a timestamp of when it was created. Created database in firebase to that creates a 'users' collection. Sign in page was created as long as user has been authenticated. Profile page will now display name of user. Still need to implement signout/logout feature.